refactor: 바텀시트 패키지 분리 및 서재뷰 & 바텀시트 UI 상태 분리 구현
